requirements.txt
eventlet
Flask
flask-cors
flasgger
google-cloud-dialogflow
requests
